+++ ImprovedStorageInfo/CHANGELOG.md
@@ -6,3 +6,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### [1.0.0] 2023-03-14
 
 * Initial commit
+
+### [2.0.0] 2023-03-15
+
+#### Technical
+
+* Use external assembly for logger
+* Use external assembly for translations
+
+#### Fixes
+
+* Missing display for escape pod container
+* Missing display for vehicule containers